[split] finagle-core: implement ServerBuilder in terms of StackServer.
Motivation:
Decrease barrier to entry for mux (and thriftmux) by making it
possible to use from ServerBuilder. This requires being able to
pass in a StackServer to ServerBuilder.

Modification:
The crux of the change is a new method in ServerBuilder `stack` which
allows for a StackServer to be injected.

RB_ID=313995