[split] finagle-zipkin: Include service name in all traces
Motivation:
This was prompted by a bug report from public finaglers reporting that thriftmux traces all queries as the Unknown service.
But deeper than that, the service name is tied to the RPC method which happen at different layers of the stack.  There are lots of scenarios where services trace as "Unknown"

Modifications:
Annotating the service name is broken off from annotating the RPC method.  Also, the service name is now passed in to the tracing filter because without it tracing doesn't really work.

Result:
Service names for everybody!

RB_ID=321823