[split] finagle-core: Enforce usage of c.t.finagle.util.DefaultLogger
Problem
It was recently pointed out that there are a few instances in finagle-core where
we do `Logger.getLogger("finagle")` rather than use the stock DefaultLogger. It
would be preferable to standardize on DefaultLogger, which scopes the logger to
"com.twitter.finagle".

Solution
Replace said instances with DefaultLogger.

Result
A consistent "root" scope for Finagle-internal log output.

RB_ID=326687