[split] finagle-core: Disable Monitor usage of in Netty3Listener
Problem
A recent change to ServerBuilder introduced unexpected behavior caused by
SourceTrackingMonitor being applied within Netty3Listener and exposing
ClosedChannelExceptions that previously occurred silently. This led to log
pollution and confusion among service owners.

com.twitter.finagle.netty3.ServerBridge#exceptionCaught already logs in certain
cases, so the monitor in the Netty pipeline is arguably redundant. Granted, this
inline logging goes to FINEST and WARNING, so it won't be seen by most deployed services.

Solution
We've never heard reports that the previously-set NullMonitor caused problems, so
we think it's acceptable to disable the Monitor usage within
Netty3Listener.scala. The `monitor` constructor argument is maintained for API
compatibility, but unused within the body of the class.

Result
We restore the previous behavior of not escalating Netty channel exceptions to
the SourceTrackingMonitor.

RB_ID=326933