[split] finagle-mysql: implement builder using StackClient.
Problem

Giving the new APIs a test drive by converting the mysql
client to use them. There is also demand to configure the mysql
client pool. The high watermark is still fixed to 1 until
we add transaction support in the rich api. However, for users
who aren't using transactions, it should be possible to configure
the pool.

Solution

Implemented the codec in terms of a `StackClient`. Make the DefaultPool
companion object accessible outside of the finagle package.

RB_ID=333159