[split] [finagle-http] fix HttpClientDispatcher
Problem

When reading HTTP chunks from the transport, failures did not
result in the tryDispatch promise being set. This, coupled with
a race condition in ChannelTransport caused requests to become
queued on the dispatcher apply semaphore and never written to the
transport.

Solution

The tryDispatch promise is updated on transport read failures.

Result

Requests no longer block indefinitely on the dispatch apply semaphore.

RB_ID=336297