[split] Move mysql testing out of finagle-mysql
Problem:
In our attempts to move finagle into aviary, I discovered a library for
testing finagle-mysql that's defined in the tests section.  We do not
publish finagle testing jars, so when finagle goes to jars, this library
will break.

Solution:
Moving these classes (which actually aren't used or tested by
finagle-mysql) into their own project so that they can stay in birdcage
and perhaps be moved later when the dependees are all in aviary.

RB_ID=335581