[split] [finagle-core] fix truncation in ChannelBufferBuf#slice()
Problem

The contract for slice() request that out of bounds indices
be truncated, but instead, ChannelBufferBuf#slice() throws
an exception if i > 0 and j-i > length + buf.capacity().

Solution

Check for out of bounds indices and truncate the length in
the call to buf.slice().

Result

slice() can be called with out of bounds indices.

RB_ID=338297