[split] Finagle: Remove `Stack{Server,Client}.transformed`
The initial problem was that `ServerBuilder.stack()` did not work with
com.twitter.exp.ThriftMuxServer, which was blocking further testing and
rollout of ThriftMux. The cause of this was that
`ThriftMuxServerImpl.WrappedMuxStackServer` did not override
`StackServer.transformed`, so when the default implementation of
`transformed` was invoked in ServerBuilder it broke ThriftMuxServer.

Remove the `StackServer.transformed` method and centralizing all
server-side Stack-configuration on the `configured` method.

All usages of `StackServer.transformed` are rewritten to use
`configured`. `ServerBuilder.stack()` now works with
com.twitter.exp.ThriftMuxServer, as verified by a new unit test and a
staging deploy of one of Twitter's core services.
override `StackServer.transformed`, so when the default implementation of `transformed` was invoked in ServerBuilder, it broke ThriftMuxServer.

RB_ID=337855