[split] finagle-core, thrift, mux: Clean up contexts, delimit Locals
The current code that deals with com.twitter.util.Local and
com.twitter.finagle.Context confuses a number of issues.

First, contexts are designed to pass-through -- thus, they are not
useful for inheriting client ids; instead they must be set explicitly
by clients. This change fixes finagle-thrift to always ignore a
context-provided client id, replacing it instead with whatever is
passed in to the client.

Secondly, we make sure that all dispatchers delimit Locals properly.
Without this, it is possible to "leak" locals between requests. Consider
the following scenario, for a single server dispatcher instance:

	service(req)
		sets client ids in filter
		...
	service(req)
		does not set client id in a filter (e.g. it was not provided)
		client id now leaks

this is made worse by the fact that the TTwitter *client* set the
client id local. (Why, I'm not sure; this is also fixed by this
change.)

Finally, we ensure that ClientIdContext does not pass-through. While
we use contexts to propagate client ids, they should not be passed
from server to client since, by their current definition, they
identify the closest client.

RB_ID=314891