[split] zk2: Start resolution eagerly
Problem

Some services have come to rely on the eager serverset resolution
semantics presented by zk1, since this allows resolution to proceed
concurrently with program startup. While this is racy behavior --
finicky, at best, to rely on -- these are nonetheless the facts on the
ground.

Solution

Start resolving zk2 serversets eagerly. A better longer-term solution
is to have some sort of API to indicate client readiness so that
services may instead rely on this.

RB_ID=376325