[split] Buf, Reader: remove Buf.Eof; end-of-stream is None
Problem

The reader interface is complicated substantially by using 'Buf.Eof'
as a sentinel value. It is a value that's confusing and brittle to
compare against (e.g. object equality must be used), and is just
unintuitive.

Solution

Signal EOF using an Option, where None = EOF. This is straightforward
to inspect. This also means we have to make Reader.writable() Closable
in order to signal EOF.

RB_ID=394095