[split] finagle-serversets: merge WeightedSocketAddresses with same host:port but different weight in Stabilizer
Problem

When changing the weight of a WeightedSocketAddress, Stabilizer will include both the old WeightedSocketAddress and the updated WeightedSocketAddress for one epoch. This results in an unintended increase in traffic to the host:port.

Solution

Merge WeightedSocketAddresses with the same underlying SocketAddress preferring the newer weight to the previous weight.

Result

Stabilizer provides a single WeightedSocketAddress per unique underlying SocketAddress.

RB_ID=399271