[split] Problem
I ran into a nasty stacktrace that was producing an ArrayOutofBoundsException, and I'm trying to solve the top-line services failures I was experiencing due to a probelm in the LatencyHistogram.

Solution

I modified the size of the array internally to the histogram.

Result

When the duration is >= to the range of the histogram, the .add will not produce an ArrayOutofBoundsException

RB_ID=430045