[split] finagle-core: Register StackClient with Path dest in ClientRegistry as evaluated against Dtab.base Problem
When a StackClient is parameterized with a Path destination, we register it over and over depending on local dtab overrides. This means the StackClient isn't added to the ClientRegistry until the client is used.

Solution

When a StackClient is parameterized with a Path destination, register it once with the Path evaluated and bound against the base dtab.

Result

While imperfect, ClientRegistery.expAllRegisteredClientsResolved() now works with StackClients parameterized with Path destinations, at least in the context of the base dtab.

RB_ID=440257