[split] Use Sphinx substitutions to keep versions up to date in documentation
Problem

Version numbers in documentation quickly get out of date without careful
maintenance.

Solution

We can use Sphinx's substitution mechanism to update versions automatically.
We get the correct version for free when building the Sphinx site through the
sbt-site plugin, but we need to read the version out of the SBT build file by
hand for DocBird (this is somewhat fragile, since it relies on conventions in
our build files).

Result

Version numbers in documentation are updated automatically.

RB_ID=442183