[split] avoid SOCKS proxy for both loopback and link-local addresses
Problem:

When you resolve localhost on a Mac you get 127.0.0.1, 0:0:0:0:0:0:0:1,
and fe80:0:0:0:0:0:0:1%1, all of which end up in the address set of a
Finagle client. The last is a link-local but not a loopback
address (according to InetAddr.is{Loopback,LinkLocal}Address), so the
previous hack (which avoided sending loopback traffic over SOCKS)
doesn't catch it, and depending on various things sometimes RPC calls
hang mysteriously when you use -socksProxy.

Solution:

Check for isLinkLocalAddress as well as isLoopbackAddress and avoid the
SOCKS proxy for either.

Result:

Connections to fe80:0:0:0:0:0:0:1%1 are not sent over SOCKS, reduction
in hair torn.

RB_ID=442955