[split] finagle-core: Rewrote ServiceLoader to avoid creating
Problem

We were constructing a Map[String, Seq[String]] which map interface names to a
sequence of their implementations.
This map was constructed every time.

Solution

Remove the construction of the map, and directly search the classpath for the
implementations of one specific interface.

Result

No apparent changes, just simpler code.

RB_ID=440855