[split] finagle-core: for paths starting with /#/, skip rewrites where prefix is /
Problem:

Dtabs don't provide a way to delegate the root to a Namer (a rule / =>
/$/com.twitter.FooNamer loops), so if you want to do that you need
additional configuration at the Namer level.

But this is overcomplicated. We should be able to use the Dtab as the
single way to configure how names are interpreted.

Solution:

Skip rewrites when the prefix is / and the path under consideration
begins with /#/, and make /#/className the same as /$/className. Then
the root may be delegated with / => /#/className. (The reason for /#/ is
twofold: to permit rewrites on existing /$/ paths such as /$/inet, and
to permit /$/className paths to be sent to another namer.)

Result:

Namer-level configuration is not necessary to achieve root delegation;
-dtab.add can be used instead.

...

This is an unvarnished hack but principled alternatives are considerably
more complicated.

RB_ID=449441