[split] finagle-core: decay without load in ewma
Problem

The ewma load metric does not properly decay when a nodes does not
receive load. This is especially problematic for outliers because
we advertise our load as if we were to schedule an additional
request on the node.

Solution

This patch introduces proper decay over time. Additionally, this
simplifies the code a bit by detecting peaks as part of the ewma.

RB_ID=451645