[split] finagle-core: GlobalFlag isn't caching the property value (performance)
Problem

When no value is specified on the command line, `GlobalFlag.apply` is checking every
time a system property. Under the hood, the `System.getProperty` call is using a
synchronized Map, which leads to contention.
This code is called at multiple times in the "hot" code path and it's significantly
impacting the performance (~13%).

Solution

Cache the value read from the system properties.

RB_ID=457407