[split] finagle-core: Fix breaking ABI change in SummarizingStatsReceiver + bug in computing array index
Problem

A breaking ABI change has been introduced in SummarizingStatsReceiver.
Also, there was a bug in the way we compute the index of the array for extracting the percentile.

Solution

Add the previous method which point to the new one.
Fix the bug, by using math.floor instead of math.ceil

RB_ID=459253