[split] Set content length header when appropriate
Problem

Content length is not set for requests even when we are sure there is a
payload body. Additionally, we forgot to set the content length header
for HttpResponse in the dispatcher.

Solution

Set the content length header when we are supposed to.

Result

Content length header is set for requests and responses when appropriate.

RB_ID=428089