finagle-serversets: Remove per-path histograms
Problem

finagle-serversets currently emits four histograms for each path that
it's asked to look up. For services that resolve large numbers of
serversets (namely TFE and the Wily Nameserver), this explodes into
thousands of metrics. There are no known usages of these stats for
monitoring purposes, so it's a huge waste and memory pressure burden.

Solution

Remove the per-path histograms in favor of four aggregated read- and
parse-latency histograms for entries and vectors.

Result

The number of stats emitted under the "zk2" scope is vastly
reduced. The per-path "size" and "limbo" stats are maintained, as are
those emitted by the InetResolver that is embedded within the
Zk2Resolver.

RB_ID=484443